+++ extensions/api/control/src/test/java/org/eclipse/dataspaceconnector/api/control/ClientControlCatalogApiControllerTestServiceExtension.java
@@ -43,6 +43,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.stream.Stream;
 
+import static org.mockito.Mockito.mock;
+
 class ClientControlCatalogApiControllerTestServiceExtension implements ServiceExtension {
 
     @Inject
@@ -60,7 +62,7 @@ class ClientControlCatalogApiControllerTestServiceExtension implements ServiceEx
         assetLoader = context.getService(AssetLoader.class);
         contractDefinitionStore = context.getService(ContractDefinitionStore.class);
         context.registerService(ConsumerContractNegotiationManager.class, new FakeConsumerNegotiationManager());
-        context.registerService(ContractNegotiationStore.class, new FakeContractNegotiationStore());
+        context.registerService(ContractNegotiationStore.class, mock(ContractNegotiationStore.class));
     }
 
     @Override
@@ -135,44 +137,6 @@ class ClientControlCatalogApiControllerTestServiceExtension implements ServiceEx
 
         contractDefinitionStore.save(contractDefinition1);
         contractDefinitionStore.save(contractDefinition2);
-    }
-
-    private static class FakeContractNegotiationStore implements ContractNegotiationStore {
-
-        @Override
-        public @Nullable ContractNegotiation find(String negotiationId) {
-            return null;
-        }
-
-        @Override
-        public @Nullable ContractNegotiation findForCorrelationId(String correlationId) {
-            return null;
-        }
-
-        @Override
-        public @Nullable ContractAgreement findContractAgreement(String contractId) {
-            return null;
-        }
-
-        @Override
-        public void save(ContractNegotiation negotiation) {
-
-        }
-
-        @Override
-        public void delete(String negotiationId) {
-
-        }
-
-        @Override
-        public @NotNull List<ContractNegotiation> nextForState(int state, int max) {
-            return null;
-        }
-
-        @Override
-        public Stream<ContractNegotiation> queryNegotiations(QuerySpec querySpec) {
-            throw new UnsupportedOperationException();
-        }
     }
 
     private static class FakeConsumerNegotiationManager implements ConsumerContractNegotiationManager {