+++ extensions/azure/cosmos/contract-negotiation-store-cosmos/src/main/java/org/eclipse/dataspaceconnector/contract/negotiation/store/CosmosContractNegotiationStore.java
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - add functionalities
  *
  */
 
@@ -24,6 +25,7 @@ import org.eclipse.dataspaceconnector.azure.cosmos.dialect.SqlStatement;
 import org.eclipse.dataspaceconnector.azure.cosmos.util.CosmosLeaseContext;
 import org.eclipse.dataspaceconnector.common.string.StringUtils;
 import org.eclipse.dataspaceconnector.contract.negotiation.store.model.ContractNegotiationDocument;
+import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNegotiationStore;
 import org.eclipse.dataspaceconnector.spi.contract.offer.store.ContractDefinitionStore;
@@ -37,9 +39,11 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static java.util.Optional.ofNullable;
 import static net.jodah.failsafe.Failsafe.with;
 
 /**
@@ -70,7 +74,6 @@ public class CosmosContractNegotiationStore implements ContractNegotiationStore 
         var object = with(retryPolicy).get(() -> cosmosDbApi.queryItemById(negotiationId));
         return object != null ? toNegotiation(object) : null;
     }
-
 
     @Override
     public @Nullable ContractNegotiation findForCorrelationId(String correlationId) {
@@ -116,12 +119,48 @@ public class CosmosContractNegotiationStore implements ContractNegotiationStore 
     }
 
     @Override
-    public Stream<ContractNegotiation> queryNegotiations(QuerySpec querySpec) {
+    public @NotNull Stream<ContractNegotiation> queryNegotiations(QuerySpec querySpec) {
         var statement = new SqlStatement<>(ContractNegotiationDocument.class);
-        var query = statement.where(querySpec.getFilterExpression()).offset(querySpec.getOffset()).limit(querySpec.getLimit()).orderBy(querySpec.getSortField(), querySpec.getSortOrder() == SortOrder.ASC).getQueryAsSqlQuerySpec();
+        var query = statement.where(querySpec.getFilterExpression())
+                .offset(querySpec.getOffset())
+                .limit(querySpec.getLimit())
+                .orderBy(querySpec.getSortField(), querySpec.getSortOrder() == SortOrder.ASC)
+                .getQueryAsSqlQuerySpec();
 
         var objects = with(retryPolicy).get(() -> cosmosDbApi.queryItems(query));
         return objects.map(this::toNegotiation);
+    }
+
+    @Override
+    public @NotNull Stream<ContractAgreement> getAgreementsForDefinitionId(String definitionId) {
+        var query = "SELECT * FROM c WHERE c.wrappedInstance.contractAgreement.id LIKE @agreementId";
+        var param = new SqlParameter("@agreementId", definitionId + ":%");
+
+        var spec = new SqlQuerySpec(query, param);
+        return with(retryPolicy).get(() -> cosmosDbApi.queryItems(spec))
+                .map(this::toNegotiation)
+                .map(ContractNegotiation::getContractAgreement);
+    }
+
+    @Override
+    public @NotNull Stream<ContractAgreement> queryAgreements(QuerySpec querySpec) {
+        var criteria = querySpec.getFilterExpression().stream()
+                .map(it -> it.withLeftOperand(op -> "contractAgreement." + op))
+                .collect(Collectors.toList());
+
+        var sortField = ofNullable(querySpec.getSortField()).map(it -> "contractAgreement." + it).orElse(null);
+
+        var query = new SqlStatement<>(ContractNegotiationDocument.class)
+                .where(criteria)
+                .offset(querySpec.getOffset())
+                .limit(querySpec.getLimit())
+                .orderBy(sortField, querySpec.getSortOrder() == SortOrder.ASC)
+                .getQueryAsSqlQuerySpec();
+
+        return with(retryPolicy).get(() -> cosmosDbApi.queryItems(query))
+                .map(this::toNegotiation)
+                .map(ContractNegotiation::getContractAgreement)
+                .filter(Objects::nonNull);
     }
 
     @Override
@@ -138,9 +177,15 @@ public class CosmosContractNegotiationStore implements ContractNegotiationStore 
         return list.stream().map(this::toNegotiation).collect(Collectors.toList());
     }
 
+    @Override
+    public Policy findPolicyForContract(String contractId) {
+        return ofNullable(findContractAgreement(contractId)).map(ContractAgreement::getPolicy).orElse(null);
+    }
+
     private ContractNegotiation toNegotiation(Object object) {
         var json = typeManager.writeValueAsString(object);
-        return typeManager.readValue(json, ContractNegotiationDocument.class).getWrappedInstance();
+        var document = typeManager.readValue(json, ContractNegotiationDocument.class);
+        return document.getWrappedInstance();
     }
 }