+++ extensions/in-memory/negotiation-store-memory/src/main/java/org/eclipse/dataspaceconnector/negotiation/store/memory/InMemoryContractNegotiationStore.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2020, 2021 Microsoft Corporation
+ *  Copyright (c) 2020 - 2022 Microsoft Corporation
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -9,12 +9,14 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - add functionalities
  *
  */
 
 package org.eclipse.dataspaceconnector.negotiation.store.memory;
 
 import org.eclipse.dataspaceconnector.common.concurrency.LockManager;
+import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNegotiationStore;
 import org.eclipse.dataspaceconnector.spi.query.QueryResolver;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
@@ -30,10 +32,12 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.toList;
 
 /**
@@ -42,20 +46,21 @@ import static java.util.stream.Collectors.toList;
  */
 public class InMemoryContractNegotiationStore implements ContractNegotiationStore {
     private final LockManager lockManager = new LockManager(new ReentrantReadWriteLock());
-    private final Map<String, ContractNegotiation> processesById = new HashMap<>();
-    private final Map<String, ContractNegotiation> processesByCorrelationId = new HashMap<>();
+    private final Map<String, ContractNegotiation> negotiationById = new HashMap<>();
+    private final Map<String, ContractNegotiation> negotiationByCorrelationId = new HashMap<>();
     private final Map<String, ContractNegotiation> contractAgreements = new HashMap<>();
     private final Map<Integer, List<ContractNegotiation>> stateCache = new HashMap<>();
-    private final QueryResolver<ContractNegotiation> queryResolver = new ReflectionBasedQueryResolver<>(ContractNegotiation.class);
+    private final QueryResolver<ContractNegotiation> negotiationQueryResolver = new ReflectionBasedQueryResolver<>(ContractNegotiation.class);
+    private final QueryResolver<ContractAgreement> agreementQueryResolver = new ReflectionBasedQueryResolver<>(ContractAgreement.class);
 
     @Override
     public ContractNegotiation find(String id) {
-        return lockManager.readLock(() -> processesById.get(id));
+        return lockManager.readLock(() -> negotiationById.get(id));
     }
 
     @Override
     public @Nullable ContractNegotiation findForCorrelationId(String correlationId) {
-        var process = processesByCorrelationId.get(correlationId);
+        var process = negotiationByCorrelationId.get(correlationId);
         var processId = process != null ? process.getId() : null;
         return find(processId);
     }
@@ -72,8 +77,8 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
             negotiation.updateStateTimestamp();
             delete(negotiation.getId());
             ContractNegotiation internalCopy = negotiation.copy();
-            processesById.put(negotiation.getId(), internalCopy);
-            processesByCorrelationId.put(negotiation.getCorrelationId(), internalCopy);
+            negotiationById.put(negotiation.getId(), internalCopy);
+            negotiationByCorrelationId.put(negotiation.getCorrelationId(), internalCopy);
             var agreement = internalCopy.getContractAgreement();
             if (agreement != null) {
                 contractAgreements.put(agreement.getId(), internalCopy);
@@ -86,7 +91,7 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
     @Override
     public void delete(String processId) {
         lockManager.writeLock(() -> {
-            ContractNegotiation process = processesById.remove(processId);
+            ContractNegotiation process = negotiationById.remove(processId);
             if (process != null) {
                 var tempCache = new HashMap<Integer, List<ContractNegotiation>>();
                 stateCache.forEach((key, value) -> {
@@ -95,7 +100,7 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
                 });
                 stateCache.clear();
                 stateCache.putAll(tempCache);
-                processesByCorrelationId.remove(process.getCorrelationId());
+                negotiationByCorrelationId.remove(process.getCorrelationId());
 
                 if (process.getContractAgreement() != null) {
                     contractAgreements.remove(process.getContractAgreement().getId());
@@ -103,6 +108,21 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
             }
             return null;
         });
+    }
+
+    @Override
+    public @NotNull Stream<ContractNegotiation> queryNegotiations(QuerySpec querySpec) {
+        return lockManager.readLock(() -> negotiationQueryResolver.query(negotiationById.values().stream(), querySpec));
+    }
+
+    @Override
+    public @NotNull Stream<ContractAgreement> getAgreementsForDefinitionId(String definitionId) {
+        return lockManager.readLock(() -> getAgreements().filter(it -> it.getId().startsWith(definitionId + ":")));
+    }
+
+    @Override
+    public @NotNull Stream<ContractAgreement> queryAgreements(QuerySpec querySpec) {
+        return lockManager.readLock(() -> agreementQueryResolver.query(getAgreements(), querySpec));
     }
 
     @Override
@@ -129,8 +149,15 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
     }
 
     @Override
-    public Stream<ContractNegotiation> queryNegotiations(QuerySpec querySpec) {
-        return lockManager.readLock(() -> queryResolver.query(processesById.values().stream(), querySpec));
+    public Policy findPolicyForContract(String contractId) {
+        return ofNullable(findContractAgreement(contractId)).map(ContractAgreement::getPolicy).orElse(null);
+    }
+
+    @NotNull
+    private Stream<ContractAgreement> getAgreements() {
+        return negotiationById.values().stream()
+                .map(ContractNegotiation::getContractAgreement)
+                .filter(Objects::nonNull);
     }
 
 }