+++ extensions/sql/contract-definition/store/src/main/java/org/eclipse/dataspaceconnector/sql/contractdefinition/store/SqlContractDefinitionStoreExtension.java
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Daimler TSS GmbH - Initial API and Implementation
+ *       Microsoft Corporation - refactoring
  *
  */
 
@@ -16,6 +17,7 @@ package org.eclipse.dataspaceconnector.sql.contractdefinition.store;
 
 
 import org.eclipse.dataspaceconnector.dataloading.ContractDefinitionLoader;
+import org.eclipse.dataspaceconnector.spi.EdcSetting;
 import org.eclipse.dataspaceconnector.spi.contract.offer.store.ContractDefinitionStore;
 import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.Provides;
@@ -24,8 +26,14 @@ import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.eclipse.dataspaceconnector.spi.transaction.TransactionContext;
 import org.eclipse.dataspaceconnector.spi.transaction.datasource.DataSourceRegistry;
 
-@Provides({ContractDefinitionStore.class, ContractDefinitionLoader.class})
-public class SqlContractDefinitionStoreServiceExtension implements ServiceExtension {
+@Provides({ ContractDefinitionStore.class, ContractDefinitionLoader.class })
+public class SqlContractDefinitionStoreExtension implements ServiceExtension {
+
+    /**
+     * Name of the datasource to use for accessing contract definitions.
+     */
+    @EdcSetting(required = true)
+    private static final String DATASOURCE_SETTING_NAME = "edc.datasource.contractdefinition.name";
 
     @Inject
     private DataSourceRegistry dataSourceRegistry;
@@ -33,13 +41,21 @@ public class SqlContractDefinitionStoreServiceExtension implements ServiceExtens
     @Inject
     private TransactionContext transactionContext;
 
+    @Inject(required = false)
+    private ContractDefinitionStatements statements;
+
     @Override
     public void initialize(ServiceExtensionContext context) {
-        var dataSourceName = context.getConfig().getString(ConfigurationKeys.DATASOURCE_SETTING_NAME);
+        var dataSourceName = context.getConfig().getString(DATASOURCE_SETTING_NAME);
 
-        var sqlContractDefinitionStore = new SqlContractDefinitionStore(dataSourceRegistry, dataSourceName, transactionContext, context.getTypeManager().getMapper());
+        var sqlContractDefinitionStore = new SqlContractDefinitionStore(dataSourceRegistry, dataSourceName, transactionContext, getStatementImpl(), context.getTypeManager().getMapper());
 
         context.registerService(ContractDefinitionLoader.class, sqlContractDefinitionStore::save);
         context.registerService(ContractDefinitionStore.class, sqlContractDefinitionStore);
     }
+
+    private ContractDefinitionStatements getStatementImpl() {
+        return statements == null ? new PostgresStatements() : statements;
+    }
+
 }