+++ spi/contract-spi/src/main/java/org/eclipse/dataspaceconnector/spi/contract/offer/store/ContractDefinitionStore.java
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Fraunhofer Institute for Software and Systems Engineering - added method
  *
  */
 
@@ -44,6 +45,14 @@ public interface ContractDefinitionStore {
      */
     @NotNull
     Stream<ContractDefinition> findAll(QuerySpec spec);
+    
+    /**
+     * Returns the definition with the given id, if it exists.
+     *
+     * @param definitionId the id.
+     * @return the definition with with the given id, or null.
+     */
+    ContractDefinition findById(String definitionId);
 
     /**
      * Persists the definitions.